Add logic for game detail view/Show/Delete/Update

+++ VideoGamesTK/Data/PersistentStorages/CoreDataStorage/EntityMapping/Game+CoreDataProperties.swift
@@ -27,7 +27,7 @@ extension Game {
     @NSManaged public var developer: String?
     @NSManaged public var freeToGameProfileURL: String?
     @NSManaged public var releaseDate: Date?
-    @NSManaged public var rate: Double
+    @NSManaged public var rate: Int32
     @NSManaged public var isFavorite: Bool
     @NSManaged public var logicDeleted: Bool
     @NSManaged public var comments: String?
